requirements.txt
mysql-connector-python==8.2.0
bcrypt==4.1.2
python-dotenv==1.0.0
tabulate==0.9.0
colorama==0.4.6
pandas==2.0.3
schedule==1.2.0
plyer==2.1.0
streamlit==1.28.1
plotly==5.17.0
streamlit-option-menu==0.3.6
streamlit-authenticator==0.2.3